backend/pyproject.toml
[project]
name = "my-rag-demo"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "langchain>=1.2.7",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.7",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.2.0",
    "langchain-text-splitters>=1.1.0",
    "openai>=2.16.0",
    "pydantic-settings>=2.7.1",
    "loguru>=0.7.3",
    "sqlalchemy>=2.0.0",
    "pymysql>=1.1.0",
    # MySQL8 默认认证插件（caching_sha2_password/sha256_password）在 PyMySQL 下需要 cryptography
    "cryptography>=42.0.0",
    "redis>=5.0.0",
    "aio-pika>=9.4.0",
    "minio>=7.2.0",
    "uvicorn>=0.40.0",
    "python-multipart>=0.0.22",
    "unstructured[docx,md,pdf]==0.18.11",
    "pymilvus[model]>=2.6.8",
    "flagembedding>=1.3.5",
    "aiomysql>=0.3.2",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.3",
]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
addopts = "-q"
# 允许直接运行 async def 测试（pytest-asyncio 提供实现）
asyncio_mode = "auto"
markers = [
    "integration: marks tests that require external services (e.g. local Ollama)",
]
